SHIR-110 # comment dev UI/BL of `MyReview`
