+++ lib/model/graphql/mixins/review_state.freezed.dart
@@ -1,0 +1,362 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+
+part of 'review_state.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+/// @nodoc
+class _$ReviewStateTearOff {
+  const _$ReviewStateTearOff();
+
+// ignore: unused_element
+  _InReview inReview() {
+    return const _InReview();
+  }
+
+// ignore: unused_element
+  _Open open() {
+    return const _Open();
+  }
+
+// ignore: unused_element
+  _Ng ng() {
+    return const _Ng();
+  }
+}
+
+/// @nodoc
+// ignore: unused_element
+const $ReviewState = _$ReviewStateTearOff();
+
+/// @nodoc
+mixin _$ReviewState {
+  @optionalTypeArgs
+  TResult when<TResult extends Object>({
+    @required TResult inReview(),
+    @required TResult open(),
+    @required TResult ng(),
+  });
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object>({
+    TResult inReview(),
+    TResult open(),
+    TResult ng(),
+    @required TResult orElse(),
+  });
+  @optionalTypeArgs
+  TResult map<TResult extends Object>({
+    @required TResult inReview(_InReview value),
+    @required TResult open(_Open value),
+    @required TResult ng(_Ng value),
+  });
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object>({
+    TResult inReview(_InReview value),
+    TResult open(_Open value),
+    TResult ng(_Ng value),
+    @required TResult orElse(),
+  });
+}
+
+/// @nodoc
+abstract class $ReviewStateCopyWith<$Res> {
+  factory $ReviewStateCopyWith(
+          ReviewState value, $Res Function(ReviewState) then) =
+      _$ReviewStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$ReviewStateCopyWithImpl<$Res> implements $ReviewStateCopyWith<$Res> {
+  _$ReviewStateCopyWithImpl(this._value, this._then);
+
+  final ReviewState _value;
+  // ignore: unused_field
+  final $Res Function(ReviewState) _then;
+}
+
+/// @nodoc
+abstract class _$InReviewCopyWith<$Res> {
+  factory _$InReviewCopyWith(_InReview value, $Res Function(_InReview) then) =
+      __$InReviewCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$InReviewCopyWithImpl<$Res> extends _$ReviewStateCopyWithImpl<$Res>
+    implements _$InReviewCopyWith<$Res> {
+  __$InReviewCopyWithImpl(_InReview _value, $Res Function(_InReview) _then)
+      : super(_value, (v) => _then(v as _InReview));
+
+  @override
+  _InReview get _value => super._value as _InReview;
+}
+
+/// @nodoc
+class _$_InReview implements _InReview {
+  const _$_InReview();
+
+  @override
+  String toString() {
+    return 'ReviewState.inReview()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _InReview);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object>({
+    @required TResult inReview(),
+    @required TResult open(),
+    @required TResult ng(),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return inReview();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object>({
+    TResult inReview(),
+    TResult open(),
+    TResult ng(),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (inReview != null) {
+      return inReview();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object>({
+    @required TResult inReview(_InReview value),
+    @required TResult open(_Open value),
+    @required TResult ng(_Ng value),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return inReview(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object>({
+    TResult inReview(_InReview value),
+    TResult open(_Open value),
+    TResult ng(_Ng value),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (inReview != null) {
+      return inReview(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class _InReview implements ReviewState {
+  const factory _InReview() = _$_InReview;
+}
+
+/// @nodoc
+abstract class _$OpenCopyWith<$Res> {
+  factory _$OpenCopyWith(_Open value, $Res Function(_Open) then) =
+      __$OpenCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$OpenCopyWithImpl<$Res> extends _$ReviewStateCopyWithImpl<$Res>
+    implements _$OpenCopyWith<$Res> {
+  __$OpenCopyWithImpl(_Open _value, $Res Function(_Open) _then)
+      : super(_value, (v) => _then(v as _Open));
+
+  @override
+  _Open get _value => super._value as _Open;
+}
+
+/// @nodoc
+class _$_Open implements _Open {
+  const _$_Open();
+
+  @override
+  String toString() {
+    return 'ReviewState.open()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _Open);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object>({
+    @required TResult inReview(),
+    @required TResult open(),
+    @required TResult ng(),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return open();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object>({
+    TResult inReview(),
+    TResult open(),
+    TResult ng(),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (open != null) {
+      return open();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object>({
+    @required TResult inReview(_InReview value),
+    @required TResult open(_Open value),
+    @required TResult ng(_Ng value),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return open(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object>({
+    TResult inReview(_InReview value),
+    TResult open(_Open value),
+    TResult ng(_Ng value),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (open != null) {
+      return open(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class _Open implements ReviewState {
+  const factory _Open() = _$_Open;
+}
+
+/// @nodoc
+abstract class _$NgCopyWith<$Res> {
+  factory _$NgCopyWith(_Ng value, $Res Function(_Ng) then) =
+      __$NgCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$NgCopyWithImpl<$Res> extends _$ReviewStateCopyWithImpl<$Res>
+    implements _$NgCopyWith<$Res> {
+  __$NgCopyWithImpl(_Ng _value, $Res Function(_Ng) _then)
+      : super(_value, (v) => _then(v as _Ng));
+
+  @override
+  _Ng get _value => super._value as _Ng;
+}
+
+/// @nodoc
+class _$_Ng implements _Ng {
+  const _$_Ng();
+
+  @override
+  String toString() {
+    return 'ReviewState.ng()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _Ng);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object>({
+    @required TResult inReview(),
+    @required TResult open(),
+    @required TResult ng(),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return ng();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object>({
+    TResult inReview(),
+    TResult open(),
+    TResult ng(),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (ng != null) {
+      return ng();
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object>({
+    @required TResult inReview(_InReview value),
+    @required TResult open(_Open value),
+    @required TResult ng(_Ng value),
+  }) {
+    assert(inReview != null);
+    assert(open != null);
+    assert(ng != null);
+    return ng(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object>({
+    TResult inReview(_InReview value),
+    TResult open(_Open value),
+    TResult ng(_Ng value),
+    @required TResult orElse(),
+  }) {
+    assert(orElse != null);
+    if (ng != null) {
+      return ng(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class _Ng implements ReviewState {
+  const factory _Ng() = _$_Ng;
+}
