+++ lib/screen_main/page_setting/account/list_tile_payment_method.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:functional_widget_annotation/functional_widget_annotation.dart';
 import 'package:shirasu/model/graphql/base_model.dart';
+import 'package:shirasu/model/graphql/viewer.dart';
 import 'package:shirasu/resource/strings.dart';
 
 part 'list_tile_payment_method.g.dart';
@@ -15,7 +16,7 @@ part 'list_tile_payment_method.g.dart';
 /// visa: "/static/img/visa-logo.png"
 @swidget
 Widget listTilePaymentMethod({
-  @required BasePaymentMethod paymentMethod,
+  @required PaymentMethod paymentMethod,
 }) =>
     ListTile(
       leading: Icon(
@@ -28,7 +29,7 @@ Widget listTilePaymentMethod({
       ),
     );
 
-extension on BasePaymentMethod {
+extension on PaymentMethod {
   IconData toCardIcon() {
     switch (brand) {
       case 'amex':
