+++ research/sample_graphql_query/post_review.graphql
@@ -1,0 +1,19 @@
+mutation PostReview($programId: ID!, $body: String!) {
+    reivew: postReview(input: {programId: $programId, body: $body}) {
+        ...UserPageReviewData
+        __typename
+    }
+}
+
+fragment UserPageReviewData on Review {
+    id
+    body
+    createdAt
+    user {
+        id
+        name
+        icon
+        __typename
+    }
+    __typename
+}
