+++ research/sample_graphql_query/program.graphql
@@ -1,4 +1,4 @@
-query GetProgram($id: ID!, $reviewsNextToken: String) {
+query GetProgram($id: ID!, $focusedReviewAuthorName: String, $reviewsNextToken: String) {
     viewer {
         name
         icon
@@ -36,6 +36,10 @@ query GetProgram($id: ID!, $reviewsNextToken: String) {
                 __typename
             }
             nextToken
+            __typename
+        }
+        focusedReview: review(authorName: $focusedReviewAuthorName) {
+            ...UserPageReviewData
             __typename
         }
         onetimePlans {
